% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artists.R
\name{getArtist}
\alias{getArtist}
\title{Get Artist Information}
\usage{
getArtist(API_KEY, mbid, Addtional_Headers = NULL, ResponseType = NULL)
}
\arguments{
\item{API_KEY}{Your Setlist FM API Key}

\item{mbid}{The Artists Musicbrainz MBID}

\item{Addtional_Headers}{Any addtional headers accepted by the api. Such as Language}

\item{ResponseType}{Whether you want JSON or XML responses (currently it is just a preprocessed list)}
}
\value{
a list object with artists mbid, unique Ticket Master Identifier (TMID), name, sortname, disambiguation, the attribution url
}
\description{
Get Artist Information
}
\examples{
getArtist(API_KEY = "KEY",mbid = 'aca5718f-4f49-4439-b8f6-209db3f11757')
SetListR::getArtist(API_KEY = "KEY",mbid = '0bfba3d3-6a04-4779-bb0a-df07df5b0558')
}
