% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Artists.R
\name{getArtistSetLists}
\alias{getArtistSetLists}
\title{Get Artists Set Lists}
\usage{
getArtistSetLists(API_KEY, mbid, page = 1, Addtional_Headers = NULL,
  ResponseType = NULL)
}
\arguments{
\item{API_KEY}{Your Setlist FM API Key}

\item{mbid}{The Artists Musicbrainz MBID}

\item{page}{The page of the site you wish to gather data from}

\item{Addtional_Headers}{Any addtional headers accepted by the api. Such as Language}

\item{ResponseType}{Whether you want JSON or XML responses (currently it is just a preprocessed list)}
}
\value{
a list object of the artists set lists for a given page
}
\description{
Get Artists Set Lists
}
\examples{
getArtistSetLists(API_KEY = "KEY",mbid = 'aca5718f-4f49-4439-b8f6-209db3f11757')
SetListR::getArtistSetLists(API_KEY = "KEY",mbid = '0bfba3d3-6a04-4779-bb0a-df07df5b0558',page = 12, ResponseType = 'json')
}
