% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search.R
\name{searchArtists}
\alias{searchArtists}
\title{Search Artists}
\usage{
searchArtists(API_KEY, mbid = NULL, artistName = NULL, page = 1,
  sort = "sortName", Addtional_Headers = NULL, ResponseType = NULL)
}
\arguments{
\item{API_KEY}{Your Setlist FM API Key}

\item{mbid}{The Artists Musicbrainz MBID}

\item{artistName}{The Artists Name}

\item{page}{e page of the site you wish to gather data from}

\item{sort}{How to sort the pages}

\item{Addtional_Headers}{Any addtional headers accepted by the api. Such as Language}

\item{ResponseType}{Whether you want JSON or XML responses (currently it is just a preprocessed list)}
}
\value{
a list of artists matching your criteria
}
\description{
Search Artists
}
\examples{
searchArtists(API_KEY = 'KEY', artistName = 'Frank Turner')
searchArtists(API_KEY = 'KEY', mbid = '0bfba3d3-6a04-4779-bb0a-df07df5b0558')
}
