% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search.R
\name{searchCities}
\alias{searchCities}
\title{Search Citiies}
\usage{
searchCities(API_KEY, cityName = NULL, cityId = NULL, countryName = NULL,
  state = NULL, stateCode = NULL, page = 1, Addtional_Headers = NULL,
  ResponseType = NULL)
}
\arguments{
\item{API_KEY}{Your Setlist FM API Key}

\item{cityName}{The CityName}

\item{cityId}{The City GeoID (https://www.census.gov/geo/reference/geoidentifiers.html)}

\item{countryName}{country name}

\item{state}{state the city lies in}

\item{stateCode}{state code the city lies in}

\item{page}{the page you want returning of the searchg results}

\item{Addtional_Headers}{Any addtional headers accepted by the api. Such as Language}

\item{ResponseType}{Whether you want JSON or XML responses (currently it is just a preprocessed list)}

\item{sort}{How to sort the pages}
}
\value{
a list of setlists matching your criteria
}
\description{
Search Citiies
}
\examples{
searchCities(API_KEY = "KEY",cityName = "Lincoln", countryName = "UK")
}
