% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search.R
\name{searchSetLists}
\alias{searchSetLists}
\title{Search Set Lists}
\usage{
searchSetLists(API_KEY, mbid = NULL, artistName = NULL, cityName = NULL,
  cityId = NULL, countryCode = NULL, page = 1, sort = "sortName",
  Addtional_Headers = NULL, ResponseType = NULL)
}
\arguments{
\item{API_KEY}{Your Setlist FM API Key}

\item{mbid}{The Artists Musicbrainz MBID}

\item{artistName}{The Artists Name}

\item{cityName}{The CityName}

\item{cityId}{The City GeoID (https://www.census.gov/geo/reference/geoidentifiers.html)}

\item{countryCode}{country code}

\item{page}{the page you want returning of the searchg results}

\item{sort}{How to sort the pages}

\item{Addtional_Headers}{Any addtional headers accepted by the api. Such as Language}

\item{ResponseType}{Whether you want JSON or XML responses (currently it is just a preprocessed list)}
}
\value{
a list of setlists matching your criteria
}
\description{
Search Set Lists
}
\examples{
searchSetLists(API_KEY = 'KEY', artistName = 'Frank Turner')
searchSetLists(API_KEY = 'KEY', mbid = '0bfba3d3-6a04-4779-bb0a-df07df5b0558')
searchSetLists(API_KEY = 'KEY', mbid = '0bfba3d3-6a04-4779-bb0a-df07df5b0558', countryCode = 'USA') #' searchSetLists(API_KEY = 'KEY', mbid = '0bfba3d3-6a04-4779-bb0a-df07df5b0558', cityName ='Lincoln')
}
